Moved new module and tests into wind calculations subdirectory.
